pyproject.toml
[tool.poetry]
name = "boids-utils"
version = "1.0.0"
description = "Utilities common to the Boids application"
authors = ["Kerry Johnson <kerry.t.johnson@gmail.com>"]
readme = "README.md"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version = "1.0.0"
version_files = [
    "pyproject.toml:^version"
]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
requests = "^2.28.2"
rich = "^13.3.1"
python-dateutil = "^2.8.2"
whistle = "^1.0.1"
elasticsearch = "^8.9.0"
backoff = "^2.2.1"
aiohttp = "^3.8.5"
colorlog = "^6.7.0"
connexion = "^2.14.2"
aiohttp-jinja2 = "^1.5.1"
yarl = "^1.9.2"
mergedeep = "^1.3.4"
nats-py = "^2.4.0"
jinja2 = "^3.1.2"

[tool.poetry.group.dev.dependencies]
commitizen = "^2.40.0"
pylint = "^3.0.1"
pre-commit = "^3.5.0"
pytest-aiohttp = "^1.0.5"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"
kafka-python = "^2.0.2"
pytest-timeout = "^2.1.0"
pytest-asyncio = "^0.21.1"
testcontainers = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
